Separated into ligand_preparators.py and dock.py. Also bug fixes.
Bug fix:
If killing python process while preparing ligand or GPU docking was being performed, it would continue. This came from using start_new_process=True in Python's subprocess.Popen, which puts the process in its own group. Therefore, when Python process was killed, that process is not a child of Python process and therefore not killed.

(Big thank you to Aws for finding this bug)